+++ lib/ui/screens/sight_card/sight_card.dart
