+++ lib/ui/screens/sight_details/sight_details.dart
