+++ lib/ui/screens/visiting_screen/visiting_screen.dart
@@ -8,8 +8,8 @@ import 'package:places/ui/res/app_colors.dart';
 import 'package:places/ui/res/app_strings.dart';
 import 'package:places/ui/res/app_typography.dart';
 import 'package:places/ui/screens/res/custom_colors.dart';
-import 'package:places/ui/screens/sight_card.dart';
-import 'package:places/ui/screens/sight_details.dart';
+import 'package:places/ui/screens/sight_card/sight_card.dart';
+import 'package:places/ui/screens/sight_details/sight_details.dart';
 import 'package:places/ui/widgets/bottom_navigation_bar.dart';
 import 'package:places/ui/widgets/sight_icons.dart';
